// clustrum-portal/src/clustrum-lib/src/shared/ui/widgets-factory/types/widget-type.ts
export enum WidgetType {
  Graph = 'graph',
  Table = 'table',
  Map = 'map',
  Card = 'card',
  // TODO: HeatMap в коде больше нигде не встречается
  HeatMap = 'heatmap',
  Text = 'text',
  Metric = 'metric',
  Ymap = 'ymap',
  Control = 'control',
  Indicator = 'indicator',
}